wait: Split methods into files by role
No code changes, will make future refactorings easier.
